requirements.txt
Pillow
streamlit
rembg